+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.1.3
+Fixed sort order for OrderedImportsFixer
+Removed `GroupImportFixer`, added modified version as `SingleImportPerLineInGroupImportFixer`
+
 ### 1.1.2
 Added multiple affirmative and negating fixers to the default sets.\
 The current set is based on `PSR12` standard with following differences: